+++ lib/views/signup.dart
@@ -1,6 +1,7 @@
 import 'package:recipe_application/helper/helperfunctions.dart';
 import 'package:recipe_application/services/auth.dart';
 import 'package:recipe_application/services/database.dart';
+import 'package:recipe_application/views/home.dart';
 import 'package:recipe_application/views/publish_recipe.dart';
 // import 'package:recipe_application/views/chatRoomsScreen.dart';
 // import 'package:recipe_application/views/signin.dart';
@@ -47,7 +48,8 @@ class _SignUpState extends State<SignUp> {
         databaseMethods.uploadUserInfo(userInfo);
         HelperFunctions.saveUserLoggedInSharedPreference(true);
         Navigator.pushReplacement(
-            context, MaterialPageRoute(builder: (context) => PublishRecipe()));
+            context, MaterialPageRoute(builder: (context) => 
+            PublishRecipe(userEmail: emailEditingController.text,userName: usernameEditingController.text,)));
       });
     }
   }
